requirements.txt
streamlit>=1.28.0
supabase>=2.0.0
yfinance>=0.2.30
pandas>=2.0.0
plotly>=5.18.0
python-dotenv>=1.0.0
requests>=2.31.0
